Eliminated use of hadoop serialization in PailTap.
